wip: omg I'm so close

+++ src/utility/Math.h
@@ -5,6 +5,6 @@
 namespace Pinetime {
   namespace Utility {
     // returns the arcsin of `arg`. asin(-32767) = -90, asin(32767) = 90
-    int16_t Asin(int16_t arg);
+    int32_t Asin(int16_t arg);
   }
 }
